% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.adcon}
\alias{process.adcon}
\title{Process ADCON data.}
\usage{
process.adcon(dirAWS, dirUP)
}
\arguments{
\item{dirAWS}{full path to the directory containing the AWS_DATA folder.\cr}

\item{dirUP}{full path to the directory containing the AWS_DATA folder on ADT server.}
}
\description{
Get the data from ADCON database, parse and convert into ADT format.
}
