% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.pulsonic}
\alias{process.pulsonic}
\title{Process ADCON data.}
\usage{
process.pulsonic(dirAWS, dirPLUSO, initData)
}
\arguments{
\item{dirAWS}{full path to the directory containing the AWS_DATA folder.\cr}

\item{dirPLUSO}{full path to the directory PULSONIC folder on ADT server.}

\item{initData}{logical, if \code{TRUE} the initialization data will be processed.}
}
\description{
Get the data from ADCON database, parse and convert into ADT format.
}
