% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcon_upload_failed_files.R
\name{upload.adcon.failed.files}
\alias{upload.adcon.failed.files}
\title{Re-upload files failed to upload.}
\usage{
upload.adcon.failed.files(dirAWS, dirUP)
}
\arguments{
\item{dirAWS}{full path to the directory containing the AWS_DATA folder.\cr}

\item{dirUP}{full path to the directory containing the AWS_DATA folder on ADT server.}
}
\description{
Re-upload all files failed to upload from the previous task to ADT server.
}
